fix ids to be Long (#14)

+++ src/main/java/com/booking/domain/Review.java
@@ -3,12 +3,12 @@ package com.booking.domain;
 import java.util.Date;
 
 public abstract class Review {
-    protected int reviewId;
+    protected Long reviewId;
     protected Date reviewDate;
     protected String description;
     protected float rating;
 
-    protected Review(int reviewId, Date reviewDate, String description, float rating) {
+    protected Review(Long reviewId, Date reviewDate, String description, float rating) {
         this.reviewId = reviewId;
         this.reviewDate = reviewDate;
         this.description = description;
